[FIXME] Skip SettleAddresses when IPv6 is set up
We use Linux 4.9, there's no support for enhanced DAD, and with hairpin mode interface in container is stuck with failed DAD on link-local address.
Without hairpin mode it takes around 30 seconds to leave tentative state for link-local address.
So both cases lead to 10 seconds for each container start, and out workload requires fast container startup.
Without this polling we may be launching containers with bad addresses, but we use IPAM and private IPv6 range + NAT, so it should not be an issue _for us_.