[FIXME] Disable DAD for IPv6 interfaces
With DAD enabled kernel tries to ensure that address is available.
That leads to no conenctivity for first 30 seconds of container life.
With optimistic DAD kernel will accept packets for container address, but will not forward any external traffic.
So we completely disable DAD, and rely on external IPAM setup for uniqueness.

Docker also disables DAD on IPv6 addresses:
https://github.com/moby/moby/blob/a77317882d010b884a9101c6ad0b2d7db141082f/libnetwork/osl/interface_linux.go#L389